+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New function overload `toString(DomainIndex)`
 
+### Changed
+
+- The payload runtime function `startPeriodicInterrupt` has been split into `setupPeriodicInterrupt` and
+  `startPeriodicInterrupt`
+
 ## 0.2 - 2023-10-30
 
 ### Added